refactor(checker): split type display recursion guard
